Move TypeScript examples to top-level examples directory
