+++ examples/http-via-sam.ts
